+++ lib/lockbox/model.rb
@@ -50,7 +50,8 @@ module Lockbox
       #   options[:type] = :float
       # end
 
-      raise ArgumentError, "Unknown type: #{options[:type]}" unless [nil, :string, :boolean, :date, :datetime, :time, :integer, :float, :binary, :json, :hash].include?(options[:type])
+      custom_type = options[:type].respond_to?(:serialize) && options[:type].respond_to?(:deserialize)
+      raise ArgumentError, "Unknown type: #{options[:type]}" unless custom_type || [nil, :string, :boolean, :date, :datetime, :time, :integer, :float, :binary, :json, :hash].include?(options[:type])
 
       attribute_type =
         case options[:type]
@@ -132,7 +133,7 @@ module Lockbox
 
                 if changes.include?(attribute)
                   type = (self.class.try(:attribute_types) || {})[attribute]
-                  if type && type.is_a?(ActiveRecord::Type::Serialized)
+                  if lockbox_attribute[:type].respond_to?(:serialize) || (type && type.is_a?(ActiveRecord::Type::Serialized))
                     send("#{attribute}=", send(attribute))
                   end
                 end
@@ -252,7 +253,9 @@ module Lockbox
                   # decrypt returns binary string
                 else
                   type = (self.class.try(:attribute_types) || {})[name.to_s]
-                  if type && type.is_a?(ActiveRecord::Type::Serialized)
+                  if custom_type
+                    message = options[:type].deserialize(message)
+                  elsif type && type.is_a?(ActiveRecord::Type::Serialized)
                     message = type.deserialize(message)
                   else
                     # default to string if not serialized
@@ -314,7 +317,9 @@ module Lockbox
                 # encrypt will convert to binary
               else
                 type = (try(:attribute_types) || {})[name.to_s]
-                if type && type.is_a?(ActiveRecord::Type::Serialized)
+                if custom_type
+                  message = options[:type].serialize(message)
+                elsif type && type.is_a?(ActiveRecord::Type::Serialized)
                   message = type.serialize(message)
                 end
               end